Remove all state logic from Steps component
I realized that the approach of marking one `StepListItem` with a `current` prop was brittle - if the `StepListItem` wasn't a direct child of `StepList`, there was no way to access the `current` prop in `StepList`.

Also, upon using the components more, I think they will be easier to reason about if they are purely graphical and completely controlled.