CopyToClipboard: support configurable labels, support plain text
Previously, you couldn't change the button label and it only supported
copying HTML. In order to support different label text, I had to remove
the fixed width styling. This styling ensured that the button wouldn't
change width when showing the "Copied" message after copying. I came up
with a new way to approximate the width of the longest label using it's
character length in em units. But this is an upper bound, since not all
characters are em width. I couldn't think of a simple way to get an
accurate width setting, so I stuck with this for now.