Optimistically render batch tally sheets button
It was a bit jarring for the button to appear later once the batch data
loaded. I figured it would be a bit nicer if the loading started in the
background while rendering the button anyway, so I added logic to wait
for the loading to finish if the user clicks the button before the batch
data has loaded.