Batch inventory: reload sign off status when files change
If a user clears and reuploads a file after signing off, we undo the sign off on the backend, since they will need to re-inventory their batches. The frontend didn't reload the sign off status when this happened, so the frontend ended up in an inconsistent state. Easy enough to fix by invalidating the query cache for sign off status.

Also added a test tabulator status file to fixtures.