Make tally entry step panel expand vertically
Even though it's nice to have a fixed height that stays above the fold, in this case, it's a bit too confusing to have the edit button out of view for long lists of candidates. While we could consider moving the edit button, this is a straightforward quick fix. Also, candidate lists shouldn't be too long, so I'm not worried about it expanding super far down the page.