Convert useRoundsAuditAdmin to react-query
My goal was to optimize the audit progress screen, which was requesting
jurisdiction status 3 times on load. The refreshing logic is quite
gnarly, and I didn't want to try to unravel it. However, I found some
logic in AuditAdminView that was trying to refresh just the rounds in
order to figure out the transition from audit setup to drawing the
sample to audit progress. By converting this to react-query (and not
relying on the existing refreshing mechanism), I was able to remove some
extra refreshing. This ended up solving the issue of loading the
jurisdictions 3 times.