Update Contests tests
The types for contests were incorrect. Contests sent from the server have some number fields. Since the contests form wanted to have empty fields, the contest form values used empty strings for those fields in their empty state. The contest type created for this form incorrectly assumed that a type with all string fields would be appropriate for all contest data in the frontend. To fix this, I made a new type for just the form values and moved the logic to convert to/from this type into the form component.