Improve data fetching/redirect on audit launch
There were two issues going on:
1. After launching the audit, the review screen would immediately
redirect to the progress screen. This redirect could occur before the rounds query updated, which would lead to a flash of the rounds screen before showing the drawing sample message.
2. When drawing the sample completed, the jurisdictions query didn't reload, meaning the progress screen wouldn't show the sampled ballots for each jurisdiction until the user refreshed.

We fix both of those issues by moving the logic for redirecting/reloading into the onSuccess callback of the rounds query.