Add vote deltas column to audit report for ballot comparison
The discrepancy column is not very human-friendly, since it shows how
the audit math interprets the difference between reported and audited
votes for a ballot. Its not intuitive how this number corresponds to the
actual difference in votes. So we add a human-friendly vote deltas
column, which simply shows the difference in votes counted for each
contest choice (if any).