Fix probely db dump migration revision stamp
The db was stamped to the wrong migration revision (one earlier than the
actual state of the db), so every time we released in Heroku and tried
to run the more recent migrations, it would error because it was trying
to create a table that already existed.

To fix this, I loaded the db dump locally, stamped it with the correct
revision, and then re-dumped it.