Improve design of review screen
A few things going on in this commit:
- Reorganized sections (and added a new Jursidictions section) to make the review screen mirror the structure of the
setup stages
- Small visual design tweaks
- Factored out the sample size selection input, since it was quite a big
chunk of code
- Reorganized the Formik form so that the sample preview button is
rendered (disabled) while the sample sizes are loading. I realized that
my initial solution from an earlier commit of moving the buttons inside
the Formik form made them not render until the sample sizes had finished
loading, which was kind of weird.
- Updates the tests accordingly, replacing snapshots with logical tests