Handle updated "contest not on ballot" case when evaluating audit results
We used to express when a contest wasn't on a ballot by not recording an
interpretation for that contest. At some point we changed to having an
explicit interpretation type, but forgot to update one piece of code
that handled this case. This function translates our interpretation into
CVR format so that it can be easily compared to the actual CVR (e.g. in
audit math or in reporting). The bug was that we were treating it like a
blank ballot, which is not too far off, so it likely slipped by
unnoticed.