Fill in vote deltas for any discrepancy
Instead of only showing vote deltas when there are both audited and
reported votes, show any mismatch between the audited result and the
reported result. This makes this column the one-stop shop for any audit
discrepancy (instead of having to also look in the discrepancy column,
which most users find confusing).