Move common code to shared module
The rounds module has historically contained a bunch of shared logic. I
had put the new vote deltas code in a new discrepancies module where I
planned to put the endpoint for the discrepancy report, but then I
realized it would make more sense to put that endpoint in the reports
module. I decided to expand the discrepancies module concept to be a
module with just code that's used in multiple API modules. Note that it
doesn't have every single piece of shared code, just a bunch of obvious
stuff. We can do further consolidation as needed.