Guard against finishing a round multiple times (#1882)
Previously, if you tried to finish a round that was already finished,
you would get a database integrity error. No harm caused, but not the
nicest UX (and causes an unnecessary alert). Instead, add a guard and a
nice error message.

Note that it's not totally clear how you would do this in the UI, but
potentially possible if you had two tabs open with the same audit and
clicked the "Finish Round" button in each one in succession.