Clear old CVR ballots from database in background during CVR file upload
We've been seeing the database query to delete CVR ballot records be
slow in production sometimes. While we're getting to the bottom of it,
this change ensures that the query is only run in the `process_cvr_file`
background task. It's not as safe as our previous approach of deleting
the records on file delete, but should be an ok temporary fix.