Move cvr ballot deletion to background task (#1892)
* Move CVR ballot record deletion to a background task

In order to safely delete CvrBallot records, ensure the db query is
always run in a background task. We can't rely on the query planner to
consistently use the appropriate indices for this query, so this is a
safe option to ensure we don't timeout any requests.

* Update workspace settings to make auto-formatting work

I think the settings format changed and it wasn't working anymore for
me.