Optimize /jurisdictions endpoint to lower risk of timeouts (#1899)
* Only count CVR ballots before audit launch

We don't show it in the UI once the audit has started, so there's no
point in calculating it. We've seen this query sometimes not use an
index and make the endpoint timeout.

* Load only the fields we use when calculating sampled ballot results

This is just a small optimization to this database query, which can be
quite a slow query. Analyzing its performance, it seems like the slowest
part is transferring all the data from the db into memory, since the
query runs quite fast when run directly from the psql prompt. There's
probably more we could do here, but these optimizations are low hanging fruit (and unfortunately have minimal impact).

* Refetch jurisdictions after undoing round start