Fetch jurisdictions discrepancy counts from separate endpoint (#1900)
Previously, discrepancy counts were included in jurisdiction round
status. This can be a slow calculation, since we need to load all the
sampled ballot interpretations. Since the jurisdictions endpoint blocks
page load, it's not great to have a slow calculation as part of that
endpoint. Instead, we load the discrepancy counts separately and show a
spinner while they are loading.