Review screen: Show only one modal for choice name standardization
We currently render a modal for each contest that may require choice
name standardization. These modals are all controlled by a single
boolean flag. That means that showing any one modal will show all of
them, which is not what we want to happen.

To fix this, I replaced the boolean flag with a contest ID for the modal
to show.