CircleCI: Cache Cypress separately
Previously, we were caching Cypress by including all of `~/.cache` in
the node cache, which only updates when `client/yarn.lock` changes. We
hypothesize that since `~/.cache` also included cached Python
dependencies, changes to the Python dependencies would be overriden by
this cache. To fix this, we cache Cypress separately and restore the
original behavior of the node cache (only caching node deps).

To be safe, I updated the cache version for all the existing caches.