Fix flaky auth code expiration test (#2015)
The current test only allowed one second until the code expired and
could be regenerated. It was being flaky and sometimes returning a new
code when we attempted to check that you couldn't generate a new code
inside that window. We need a little bit longer window to make that
check safely.