Rollback db transaction when background task is interrupted
Previously, when a task got interrupted, we let the process die
immediately, which would rollback the transaction automatically.

Now, we try to reset the task, which involves clearing the task
started_at timestamp and then commiting the current transaction to save
that change to the db. So in order to ensure that any in-progress
changes from the task itself are not committed alongside those changes,
we need to rollback the current transaction and start a new one.

I updated the tests for task resetting to cover this behavior, which
they previously didn't cover because they didn't use the db.