Ensure that completed tasks cannot be reset (#2024)
There was a race condition that allowed for tasks to be reset if the
worker was interrupted immediately after the task completed (but before
the worker set `task = None`). Instead of clearing the `task` variable,
we protect against this case by reloading the task from the db before
resetting it and ensuring that it is not completed. We also check that
it has in fact started, for good measure, though that's not strictly
necessary since resetting an unstarted task wouldn't do anything.

To test this, I ran the tests repeatedly. Before the change, this bug
would cause a failure with 10 test runs. Afterwards, it didn't fail at
all in 20 test runs.

I did catch one other small bug, which is that sometimes when exiting
the worker after resetting a task, its possible for the work loop to try
to commit a transaction even though the db connection has already been
closed. I'm not totally sure why this happens, and it's pretty
inconsequential since the worker is already exiting, but I added a guard
for it just to keep things tidy.