fix: prevent some updates that would affect incomplete jobs
Refs https://github.com/votingworks/arlo/issues/1532

When making updates we'd like to check that no incomplete background jobs are dependent on the data we're about to change. This change takes steps in that direction, but does not fully fix the issue. For example, this change does not prevent deleting files where an incomplete job references that file.