fix: prevent some updates that would affect incomplete jobs (#2043)
* fix: prevent some updates that would affect incomplete jobs

Refs https://github.com/votingworks/arlo/issues/1532

When making updates we'd like to check that no incomplete background jobs are dependent on the data we're about to change. This change takes steps in that direction, but does not fully fix the issue. For example, this change does not prevent deleting files where an incomplete job references that file.

* test: add coverage for `File::is_processing`

* style(server): reformat python files

* test(server): remove unneeded fixtures

Not all the ones that become unused parameters can be omitted due to side effects that we depend on.

* test(server): save and restore the old config value

* fix(server): correct terminology in error messages