Fix risk measurement calculation for combined batches with non-RLA batches
The current logic for combined batches allows a non-RLA batch to be the
"representative batch" for the combined batch (when using the feature
flag to sample extra batches). This means that audit results for the
combined batch are associated with that representative batch in the db.

However, we filter out non-RLA batches when querying the batch results
in the db for calculating the risk measurement to avoid accidentally
including them in the calculation. This caused an error if a non-RLA
batch was the representative batch, since the results for the entire
combined batch would be missing.

Luckily, it turns out that the risk measurement calculation uses a
different data source to determine which batches to include (a list of
batches by ticket numbers, which also filters out non-RLA batches) and
only uses the results dict as a lookup. So it's safe to include non-RLA
batches in the results dict as a short term fix.