Prioritize RLA batches when choosing representative batch
The existing logic meant that sometimes extra batches could be chosen as
representative batches. If this happened, the audited vote counts in the audit
report wouldn't include the combined batch results, since the db query
for that vote count filters out extra batches.

This made the new test added in this PR non-deterministic, since
sometimes it would choose a sampled batch as the representative, and
sometimes an extra batch.