In single jurisdiction audits, show discrepancies as they are audited (#2078)
Normally, we wait to show discrepancies until they audit boards sign off
(ballot comparison) or finalize batch results (batch comparison). This
is good in general, but in single jurisdiction audits, it means there is
no visibility into discrepancies until the audit is over. (Generally
single jurisdiction audits are used for large jurisdictions.) So we make
an exception and show the discrepancies immediately for single
jurisdiction audits.