Add filter for extra batches to sampled batch results query
We had to remove this filtering in
https://github.com/votingworks/arlo/pull/2075 to fix a bug related to
how combined batches interacted with extra sampled batches. Technically,
this filtering isn't necessary, since the audit math uses the ticket
number dict as the source of truth for which batches to include in the
p-value calculation. However, it's nice to have multiple guards in
place like we did previously rather than depend on that implementation
detail for safety.