CircleCI: Install Cypress again if cache restoration didn't work
Sometimes, it appears that the Cypress cache is restored successfully
but the actual files are not present in the filesystem as far as I can
tell. This causes the CI job to fail.

Until we can get to the bottom of what's going on, this PR implements a
workaround to check if Cypress is installed and, if not, install it.