chore(deps): update vulnerable NPM dependencies
These packages were identified as having known vulnerabilities in the
versions we were using. This change updates the packages, primarily
using the yarn `resolutions` field to specifically use a given version
of a package instead of what it would normally resolve to. Where
possible, we favor simply updating the package version, but this only
applies to direct dependencies.