chore: replace `mypy` with `basedpyright`
Updating `sentry-sdk`, as I'm doing for #2153, exposes a crashing bug in
the version of `mypy` we're currently using. It seems that it's probably
fixed in the current version, but updating to that version introduces a
lot of type errors. Rather than fixing all of them, this change switches
from `mypy` to `basedpyright` (an improved fork of the widely-used
`pyright` type checker).

To avoid having to fix all the issues it identifies right away, it
supports generating a baseline file that lists all the existing errors,
telling `basedpyright` not to complain about them. As we fix these
errors, `basedpyright` will update the file and we'll commit the updates
with the changes that caused them.