chore: replace `mypy` with `basedpyright` (#2171)
* chore: replace `mypy` with `basedpyright`

Updating `sentry-sdk`, as I'm doing for #2153, exposes a crashing bug in
the version of `mypy` we're currently using. It seems that it's probably
fixed in the current version, but updating to that version introduces a
lot of type errors. Rather than fixing all of them, this change switches
from `mypy` to `basedpyright` (an improved fork of the widely-used
`pyright` type checker).

To avoid having to fix all the issues it identifies right away, it
supports generating a baseline file that lists all the existing errors,
telling `basedpyright` not to complain about them. As we fix these
errors, `basedpyright` will update the file and we'll commit the updates
with the changes that caused them.

* chore: regenerate `poetry.lock`

* chore: update `basedpyright` baseline file

I'm not sure why this changed.

* chore: try specifying the baseline file path to enable it in CI

I'm not sure why CI is not using the file, but it's reporting errors
when they should be suppressed.

* chore: update `basedpyright` baseline file again

This is the version generated on CircleCI. I'm not sure why it's
different, and this may be why this feature is labeled as a "beta". I
may have to give up on this as a result.

* ci: fix dirty check

`git diff-index` seems to use cached information and might be stale.
Using `git status` recomputes the dirty state.