chore(deps): update/remove vulnerable NPM dependencies (#2159)
* chore(deps): update vulnerable NPM dependencies

These packages were identified as having known vulnerabilities in the
versions we were using. This change updates the packages, primarily
using the yarn `resolutions` field to specifically use a given version
of a package instead of what it would normally resolve to. Where
possible, we favor simply updating the package version, but this only
applies to direct dependencies.

* chore(deps): remove `http-proxy-middleware` package

We no longer use this since we're using `vite` to handle the proxying.
This removes 12 "high" severity vulnerabilities per `yarn audit`.

* chore: remove linux-incompatible package and re-sync lockfile

* chore(deps): update to latest `typescript`

* build: fix or suppress type errors