style: replace `black` and `pylint` with `ruff`
`ruff` is a fast linter and formatter for python. We're not heavy users
of `pylint` and don't rely on it that heavily. In part, I think this is
because it's so slow. It takes several seconds to run on just a subset
of the files in the project, slowing down commits via the `pre-commit`
hook greatly. `ruff` can lint and format in about 15ms each. `black` is
not as slow as `pylint`, but it's very slow compared to `ruff`.