Update to Python 3.11 (#2214)
* chore: upgrade to python 3.11

* chore: upgrade `psycopg2-binary` to a version compatible with python 3.11

* test: use the correct kwarg `autospec`

As far as I can tell, `auto_spec` was a typo that previous versions of
python did not complain about.

* fix: use `value` attribute of enums for unqualified name

The default display for enums changed from the unqualified name, e.g.
`VOTE` for `Interpretation.VOTE`, to the qualified name, e.g.
`Interpretation.VOTE`. To access the qualified name we now get the
`value` attribute.

* style: replace uses of deprecated types

* ci: update docker image for python 3.11

* build: update python install instructions

* refactor: break import cycle

* fix: update `consistent_sampler` revision

This new version is compatible with Python 3.10+.